% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_map_state.R
\name{fars_map_state}
\alias{fars_map_state}
\title{Title: fars_map_state
Description: year is converted to a filename with make_filename function.
Usage: fars_summarize_years(years)
Arguments: state.num &  year. Both integer values.}
\usage{
fars_map_state(state.num, year)
}
\arguments{
\item{year}{Values: DF with monthly  accidents for row and years for col}
}
\value{
fars_map_state
}
\description{
Title: fars_map_state
Description: year is converted to a filename with make_filename function.
Usage: fars_summarize_years(years)
Arguments: state.num &  year. Both integer values.
}
\examples{
\dontrun{
   fars_map_state(1, 2013)
}
}
