% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_read_years.R
\name{fars_read_years}
\alias{fars_read_years}
\title{Title: fars_read_years
Description: For each input, this function will call the make_filename function,
which produces the files name for the info in that year,
then loads the generated filename into R using the fars_read function,
and then filters the data to just month and years as the cols.
If a file with a year doesn't exist, an error message is produced.
Usage: fars_read_years(years)
Arguments: years. the years to find files to bring into R. Can be single or multiple values.}
\usage{
fars_read_years(years)
}
\arguments{
\item{years}{Value: This function returns the dataframes for files that exist for the years inputted. message of error will occur if it does not exist.}
}
\value{
fars_read_years
}
\description{
Title: fars_read_years
Description: For each input, this function will call the make_filename function,
which produces the files name for the info in that year,
then loads the generated filename into R using the fars_read function,
and then filters the data to just month and years as the cols.
If a file with a year doesn't exist, an error message is produced.
Usage: fars_read_years(years)
Arguments: years. the years to find files to bring into R. Can be single or multiple values.
}
\examples{
\dontrun{
fars_read_years(years=c(2013,2015))
fars_read_years(c(2014,2015)) }
}
