% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fars_summarize_read.R
\name{fars_summarize_years}
\alias{fars_summarize_years}
\title{Title: fars_summarize_read
Description: The data for the input is read using fars_read_years function
The dataframes are then unioned using a bind_rows function from dplyr.
Produced df is grouped by the year and MONTH.
The total number of records is then summarized with summarize function
The results are represented with months as rows and years
as columns using the spread fnction from tidyr.
An error message pops up if file does not exist
Usage: fars_summarize_years(years)
Arguments: years.}
\usage{
fars_summarize_years(years)
}
\arguments{
\item{years}{Value: A data frame with crashes for each year and month. An error message occurs if file or input does not exist.}
}
\value{
fars_summarize_years
}
\description{
Title: fars_summarize_read
Description: The data for the input is read using fars_read_years function
The dataframes are then unioned using a bind_rows function from dplyr.
Produced df is grouped by the year and MONTH.
The total number of records is then summarized with summarize function
The results are represented with months as rows and years
as columns using the spread fnction from tidyr.
An error message pops up if file does not exist
Usage: fars_summarize_years(years)
Arguments: years.
}
\examples{
\dontrun{
   fars_summarize_years(2013)
}
}
