% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_filename.R
\name{make_filename}
\alias{make_filename}
\title{Title: make_filename
Description: This is a function that takes an input in the form of a four digit value, a year, such as 2017.
The function then converts the input to an integer and holds it in a variable called year
It then takes the year variable and applies it to the string accident_%d.csv.bz2 as accident_2017.csv.bz2
when the function is called for the year 2017.
Usage: make_filename(year)
Arguments: year. A string or integer providing the correct year to use in the file name}
\usage{
make_filename(year)
}
\arguments{
\item{year}{Value: This function produces the filename a particular year. A warning message comes if the input cannot be converted to an integer.}
}
\value{
make_filename
}
\description{
Title: make_filename
Description: This is a function that takes an input in the form of a four digit value, a year, such as 2017.
The function then converts the input to an integer and holds it in a variable called year
It then takes the year variable and applies it to the string accident_%d.csv.bz2 as accident_2017.csv.bz2
when the function is called for the year 2017.
Usage: make_filename(year)
Arguments: year. A string or integer providing the correct year to use in the file name
}
\examples{
\dontrun{
   make_filename(2017)
}
}
